catch invalid via nodes in restriction parsing, set default cucumber origin to guinée

+++ include/util/timezones.hpp
@@ -12,7 +12,7 @@
 // local_time_t is a pair of a time zone shape polygon and the corresponding local time
 // rtree_t is a lookup R-tree that maps a geographic point to an index in a local_time_t vector
 using point_t = boost::geometry::model::
-    point<int32_t, 2, boost::geometry::cs::spherical_equatorial<boost::geometry::degree>>;
+    point<double, 2, boost::geometry::cs::spherical_equatorial<boost::geometry::degree>>;
 using polygon_t = boost::geometry::model::polygon<point_t>;
 using box_t = boost::geometry::model::box<point_t>;
 using rtree_t =
