+++ src/updater/updater.cpp
@@ -463,7 +463,7 @@ bool IsRestrictionValid(const Timezoner &tz_handler,
 
     // Get local time of the restriction
     const auto &local_time =
-        tz_handler.GetLocalTime(point_t{static_cast<int>(lon), static_cast<int>(lat)});
+        tz_handler.GetLocalTime(point_t{lon, lat});
 
     // TODO: check restriction type [:<transportation mode>][:<direction>]
     // http://wiki.openstreetmap.org/wiki/Conditional_restrictions#Tagging
